完成餐厅和菜品的重构

+++ src/main/java/cn/edu/szu/Bajie/entity/CanteenUrl.java
@@ -25,7 +25,7 @@ public class CanteenUrl implements Serializable {
      * 餐厅id
      */
     @TableField(value = "canteen_id")
-    private Integer canteenId;
+    private Long canteenId;
 
     /**
      * url
